+++ src/core/cache/restricted/spec.js
@@ -60,4 +60,31 @@ describe('core/cache/restricted', () => {
 
 		expect(cache.clear((el) => el > 1)).toEqual(new Map([['bar', 2]]));
 	});
+
+	it('modify size of cache', () => {
+		const cache = new RestrictedCache(1);
+
+		cache.set('foo', 1);
+
+		cache.modifySize(1);
+		cache.set('bar', 2);
+
+		expect(cache.has('foo')).toBeTrue();
+		expect(cache.has('bar')).toBeTrue();
+
+		expect(cache.modifySize(-2)).toEqual(new Map([['foo', 1], ['bar', 2]]));
+
+		expect(cache.has('foo')).toBeFalse();
+		expect(cache.has('bar')).toBeTrue();
+	});
+
+	it('modify size of cache with excess', () => {
+		const cache = new RestrictedCache(1);
+
+		cache.set('foo', 1);
+
+		cache.modifySize(-1000);
+
+		expect(cache.has('foo')).toBeFalse();
+	});
 });